+++ app/src/main/java/com/example/trove_mobile/Home/ImageSliderAdapter.java
@@ -8,7 +8,7 @@ import android.widget.ImageView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 import com.bumptech.glide.Glide;
-import com.example.mobile_app_client.R;
+import com.example.trove_mobile.R;
 
 import java.util.List;
