+++ app/src/main/java/com/example/trove_mobile/orderDetails/OrderedProductAdapter.java
@@ -14,7 +14,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.mobile_app_client.R;
+import com.example.trove_mobile.R;
 import com.example.trove_mobile.product.Product;
 
 import java.util.List;
