+++ app/src/main/java/com/example/trove_mobile/payment/PaymentMethodsFragment.java
@@ -16,7 +16,7 @@ import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.Toast;
 
-import com.example.mobile_app_client.R;
+import com.example.trove_mobile.R;
 
 /**
  * Fragment for adding or changing payment methods.
