+++ app/src/main/java/com/example/trove_mobile/reviews/ReviewAdapter.java
@@ -7,7 +7,7 @@ import android.widget.RatingBar;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
-import com.example.mobile_app_client.R;
+import com.example.trove_mobile.R;
 import java.util.List;
 
 public class ReviewAdapter extends RecyclerView.Adapter<ReviewAdapter.ViewHolder> {
